Improve small sort performance on CUDA
Currently, `bitonicSortKVInPlace` is written to sort one array per
block of threads. If that dimension happens to be very small
(<128 elements), this results in low thread occupancy.

Instead, this changes `bitonicSortKVInPlace` to operate with a 2d
block. Sorting happens along the x dimension, and the y dimension
is a fixed size batch.

ghstack-source-id: 3528cd7e4a42b79d09ec3dbbcdff415d99678c3d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79627