Fix backward of binary_cross_entropy_with_logits
The previous PR in this stack uncovered an error in the forward over
backward for this function.

In this PR, we fix this error and we also fix the gradgrad
implementation (and make it more stable and faster using `logsigmoid`).
We also move the double backward for this function to `FunctoinsManual`
as there's no reason for it to be in `native_functions`

ghstack-source-id: f2afb43e4e154c0f91a23b45c116d53df8e6494a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79381