Implemnet forward AD for slogdet-related operations and fix their complex derivative
The derivative for complex inputs of slogdet was incorrect as it said
that the output `sign` was not differentiable. Same happened with
`logdet` which didn't even support complex inputs. Now these derivatives
are all correct and all the gradcheck tests pass.

We also add forward AD formulas for these functions.

We take this chance to announce in the docs that the derivatives of
these functions are much more better behaved than those from
`linalg.det`, so one should probably always use these functions over
`linalg.det` unless they have very good reasons not to do so.

ghstack-source-id: 68d0108c1638f8209fc3fd210e960fbfea4aa469
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79743